Team Selection/Creation improvements

+++ Dots-RTS-Development/Assets/Scripts/Core Classes/SaveData.cs
@@ -16,7 +16,7 @@ public class SaveData {
 public class SaveDataCampaign {
     public SaveData game;
     public string preview;
-    public int timeUnformated;
+    public float timeUnformated;
     public bool isCleared;
 }
